// src/components/nav-section/type.ts
import { ReactElement } from 'react';
import { BoxProps } from '@mui/material';

// ----------------------------------------------------------------------

export type NavListProps = {
  title: string;
  path: string;
  icon?: ReactElement;
  info?: ReactElement;
  children?: {
    title: string;
    path: string;
    children?: { title: string; path: string }[];
  }[];
};

export type NavItemProps = {
  item: NavListProps;
  isCollapse?: boolean;
  active?: boolean | undefined;
  open?: boolean;
  onOpen?: VoidFunction;
  onMouseEnter?: VoidFunction;
  onMouseLeave?: VoidFunction;
};

export interface NavSectionProps extends BoxProps {
  isCollapse?: boolean;
  navConfig: {
    subheader?: string;
    items: NavListProps[];
  }[];
  navClientConfig?: {
    subheader?: string;
    items: NavListProps[];
  }[];
  navStorageConfig?: {
    subheader?: string;
    items: NavListProps[];
  }[];
  navProfileConfig?: {
    subheader?: string;
    items: NavListProps[];
  }[];
  navManagementConfig?: {
    subheader?: string;
    items: NavListProps[];
  }[];
 
}